+++ CHANGELOG.md
@@ -1,9 +1,11 @@
-1.0.0-M4
+1.0.0-RC1
 ========
  - Added `bits` and `bytes` codecs that behave like `bits(size)`/`bytes(size)` but with no size constraint
  - Replaced discriminator support with new implementation that is much more general (thanks to Paul C)
  - Removed `Codec.{ encode, decode }` overloads that aliased `Codec#{ encode, decode }`
  - Introduced `encodeValid`, `decodeValue`, and `decodeValidValue` methods on both the `Codec` class and object
+ - Added new combinators (`lazily`, `complete`, `compact`)
+ - API docs
 
 1.0.0-M3
 ========